refactor(cli): move to dir, add fixtures
