+++ src/cli/fixtures/apply-preset-from-package.js
@@ -1,0 +1,6 @@
+module.exports = {
+  cli:
+    "--file ./src/mocks/petstore-v3-short.json --presets demo3-swagger-to-js-preset",
+  fileCode: "from-demo.js",
+  fileTypes: "from-demo.d.ts",
+};
