+++ src/cli/fixtures/change-code-and-types-file-name.js
@@ -1,0 +1,6 @@
+module.exports = {
+  cli:
+    "--file ./src/mocks/petstore-v3-short.json --template-file-name-types demo.d.ts --template-file-name-code demo.js",
+  fileCode: "demo.js",
+  fileTypes: "demo.d.ts",
+};
