+++ src/cli/fixtures/change-code-file-name.js
@@ -1,0 +1,6 @@
+module.exports = {
+  cli:
+    "--file ./src/mocks/petstore-v3-short.json --template-file-name-code demo.ts",
+  fileCode: "demo.ts",
+  fileTypes: "index.d.ts",
+};
