+++ src/cli/fixtures/change-types-file-name.js
@@ -1,0 +1,6 @@
+module.exports = {
+  cli:
+    "--file ./src/mocks/petstore-v3-short.json --template-file-name-types demo.d.ts",
+  fileCode: "index.js",
+  fileTypes: "demo.d.ts",
+};
