+++ src/cli/fixtures/disable-request-importing-or-generating.js
@@ -1,0 +1,5 @@
+module.exports = {
+  cli: "--file ./src/mocks/petstore-v3-short.json --import-request-disabled",
+  fileTypes: "index.d.ts",
+  fileCode: "index.js",
+};
