+++ src/cli/fixtures/disable-types-generating.js
@@ -1,0 +1,4 @@
+module.exports = {
+  cli: "--file ./src/mocks/petstore-v3-short.json --disable-types-generate",
+  fileCode: "index.js",
+};
