+++ src/cli/fixtures/load-config.js
@@ -1,0 +1,14 @@
+module.exports = {
+  cli: "--file ./src/mocks/petstore-v3-short.json",
+  fileCode: "swagger-petstore-short-open-api-3-0.js",
+  fileTypes: "swagger-petstore-short-open-api-3-0.d.ts",
+  configContent: `
+    module.exports = {
+      file: "./src/mocks/petstore-v3-short.json",
+      outputDir: "./TEST_API",
+      importRequest: "disabled",
+      templateFileNameCode: ({ swaggerData, changeCase }) => \`\${changeCase.paramCase(swaggerData.info.title)}.js\`,
+      templateFileNameTypes: ({ swaggerData, changeCase }) => \`\${changeCase.paramCase(swaggerData.info.title)}.d.ts\`,
+    }
+  `,
+};
