+++ src/cli/fixtures/works-with-v2-json.js
@@ -1,0 +1,5 @@
+module.exports = {
+  cli: "--file ./src/mocks/petstore-v2.json",
+  fileTypes: "index.d.ts",
+  fileCode: "index.js",
+};
