+++ src/cli/fixtures/works-with-v2-yaml.js
@@ -1,0 +1,5 @@
+module.exports = {
+  cli: "--file ./src/mocks/petstore-v2.yaml",
+  fileTypes: "index.d.ts",
+  fileCode: "index.js",
+};
