+++ src/cli/fixtures/works-with-v3-json.js
@@ -1,0 +1,5 @@
+module.exports = {
+  cli: "--file ./src/mocks/petstore-v3.json",
+  fileTypes: "index.d.ts",
+  fileCode: "index.js",
+};
