+++ src/cli/fixtures/works-with-v3-yaml.js
@@ -1,0 +1,5 @@
+module.exports = {
+  cli: "--file ./src/mocks/petstore-v3.yaml",
+  fileTypes: "index.d.ts",
+  fileCode: "index.js",
+};
